Add conanfile with external library info
The conan package manager provides an easy way to
make all external dependencies available across
multible platforms.

This commit adds the conan configuration file for the
project with all dependencies that are required by
the sdfont library and the corresponding executables.

Note that the config file is setup to work with CMake,
which will be added in a future commit.